+++ frontend/admin/src/pages/Pneumonia.jsx
@@ -6,7 +6,7 @@ const Pneumonia = () => {
 
     return (
         <PneumoniaPredictionForm inferenceAPI={"http://localhost:8000/pneumonia/predict"} reportAPI={"http://localhost:3000/pneuomania-report/create"} />
-    );
+    ); //TODO: CHANGE API URL
 };
 
 export default Pneumonia;
