+++ frontend/user/src/pages/EHR.jsx
@@ -12,7 +12,7 @@ const EHR = () => {
     useEffect(() => {
         const fetchData = async () => {
             try {
-                const response = await fetch('http://localhost:3000/ehr/get', {
+                const response = await fetch('http://localhost:3000/ehr/get', { //TODO: CHANGE API URL
                     headers: {
                         Authorization: `Bearer ${auth.token}`
                     }
@@ -33,7 +33,7 @@ const EHR = () => {
     const onDelete = async (id) => {
         if (confirm('Are you sure you want to delete this record?')) {
             try {
-                const response = await fetch(`http://localhost:3000/ehr/delete/${id}`, {
+                const response = await fetch(`http://localhost:3000/ehr/delete/${id}`, { //TODO: CHANGE API URL
                     method: 'DELETE',
                     headers: {
                         Authorization: `Bearer ${auth.token}`,
