+++ frontend/user/src/pages/EditSchedule.jsx
@@ -9,7 +9,7 @@ const EditSchedule = () => {
   const [data, setData] = useState({ title: "", content: "" });
 
   const handleSave = async () => {
-    const response = await fetch(`http://localhost:3000/schedule/edit/${id}`, {
+    const response = await fetch(`http://localhost:3000/schedule/edit/${id}`, { //TODO: CHANGE API URL
       method: "PUT", // Or appropriate HTTP method
       headers: {
         "Content-Type": "application/json",
